TextHasher added
Added support for hashing text files to calcHash function in FileInfo file. Added TextHasher file, which looks through the current directory and searches for files with specified extensions. It then calculates a SHA-256 hash for each file and outputs it to the console. You can change the current directory where to look for files using new File("path_to_directory").