PhotoHashing added
Changed pom.xml file to make PhotoHashing.java work Added PhotoHasher class to compute image hash using pHash algorithm. Implemented the calculatePHashPhoto method to perform all steps of the algorithm, including resizing, discrete cosine transform (DCT), and hash construction. Image processing is now supported in the calcHash method of the HashCalculator class depending on the file type.
Added DCT class to perform Discrete Cosine Transform (DCT), required for pHash algorithm in PhotoHasher. The process and applyDCT methods that perform DCT on an image are implemented. The class also implements the MarvinImagePlugin interface for integration with the Marvin Framework. This class is used in the process of calculating the hash of photos in PhotoHasher.
Added case Photo in FileInfo.java file